+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix incorrect conversion of I2C address during Atlas pH sensor calibration ([#425](https://github.com/kizniche/mycodo/issues/425))
  - Potential fix for ADC issues when using pre-output ([#418](https://github.com/kizniche/mycodo/issues/418))
+ - Fix Linux Command measurement display on lines 2 through 4 of LCDs ([#427](https://github.com/kizniche/mycodo/issues/427))
 
 
 ## 5.6.5 (2018-03-14)